requirements.txt
Django
psycopg2-binary
Pillow
djangorestframework
python-decouple
celery
drf-yasg
redis
django-modeltranslation
bs4
requests    

